fix(ui): migrate @tauri-apps/api/tauri to @tauri-apps/api/core
