+++ cotoami_desktop/ui/src/main/scala/marubinotto/libs/tauri/dialog.scala
@@ -10,15 +10,15 @@ object dialog extends js.Object {
 
   /** Open a file/directory selection dialog.
     *
-    * The selected paths are added to the filesystem and asset protocol
-    * allowlist scopes. When security is more important than the easy of use of
-    * this API, prefer writing a dedicated command instead.
+    * The selected paths are added to the filesystem and asset protocol scopes.
+    * When security is more important than the easy of use of this API, prefer
+    * writing a dedicated command instead.
     *
-    * Note that the allowlist scope change is not persisted, so the values are
-    * cleared when the application is restarted. You can save it to the
-    * filesystem using tauri-plugin-persisted-scope.
+    * Note that the scope change is not persisted, so the values are cleared
+    * when the application is restarted. You can save it to the filesystem using
+    * tauri-plugin-persisted-scope.
     *
-    * <https://tauri.app/v1/api/js/dialog/#open>
+    * https://v2.tauri.app/reference/javascript/dialog/#open
     *
     * @param options
     * @return
